Implement node draining before acknowledging reservations
When a client requests a slot, the server tries to match it to a node in a
kubernetes cluster and drain the node before allowing the client to continue.
Same thing before releasing a reservation, the server will try to uncordon the
node.

Moves kubeconfig from kubernetes storage configuration to be a global setting
instead.

Signed-off-by: Heathcliff <heathcliff@heathcliff.eu>