Add scripts for generating manifests
Add a script for generating the kubernetes manifest from a template.
This allows for adding manifest files for specific release tags to the github
releases.
To ensure this is automated, add a release workflow as well.
To check if the example manifest is up to date, add it to the validate script.

Signed-off-by: Heathcliff <heathcliff@heathcliff.eu>